refactor: extract getPoolTokens
