+++ lib/modules/pool/pool.helpers.ts
@@ -9,6 +9,7 @@ import {
   GqlPoolStakingOtherGauge,
   GqlPoolTokenDetail,
   GqlPoolType,
+  GqlToken,
 } from '@/lib/shared/services/api/generated/graphql'
 import { isSameAddress } from '@/lib/shared/utils/addresses'
 import { Numberish, bn } from '@/lib/shared/utils/numbers'
@@ -23,6 +24,8 @@ import { getUserTotalBalanceInt } from './user-balance.helpers'
 import { dateToUnixTimestamp } from '@/lib/shared/utils/time'
 import { balancerV2VaultAbi } from '../web3/contracts/abi/generated'
 import { balancerV3VaultAbi } from '../web3/contracts/abi/balancerV3VaultAbi'
+import { supportsNestedActions } from './actions/LiquidityActionHelpers'
+import { getLeafTokens } from '../tokens/token.helpers'
 
 /**
  * METHODS
@@ -322,4 +325,24 @@ export function requiresPermit2Approval(pool: Pool): boolean {
 
 export function getRateProviderWarnings(warnings: string[]) {
   return warnings.filter(warning => !isEmpty(warning))
+}
+
+export function getPoolTokens(
+  pool: Pool,
+  getToken: (address: string, chain: GqlChain) => GqlToken | undefined
+): GqlToken[] {
+  type PoolToken = Pool['poolTokens'][0]
+  function toGqlTokens(tokens: PoolToken[]): GqlToken[] {
+    return tokens
+      .map(token => getToken(token.address, pool.chain))
+      .filter((token): token is GqlToken => token !== undefined)
+  }
+
+  // TODO add exception for composable pools where we can allow adding
+  // liquidity with nested tokens
+  if (supportsNestedActions(pool)) {
+    toGqlTokens(getLeafTokens(pool.poolTokens))
+  }
+
+  return toGqlTokens(pool.poolTokens)
 }
